Move all files in constants dir to constants
