--- projects/jdwp/constants/step_size.py
@@ -1,10 +1,0 @@
-"""Step size enum class."""
-
-from enum import Enum
-
-
-class StepSize(Enum):
-    """StepSize constants for JDWP."""
-
-    MIN = 0
-    LINE = 1
