--- projects/jdwp/constants/suspend_status.py
@@ -1,9 +1,0 @@
-"""Suspend status enum class."""
-
-from enum import Enum
-
-
-class SuspendStatus(Enum):
-    """SuspendStatus constants for JDWP."""
-
-    SUSPEND_STATUS_SUSPENDED = 0x1
